refactor: refactor update-content methods to offer "automatic" and "manual" version for other modules to use
